requirements.txt
fastapi
uvicorn
sentence-transformers
aio-pika>=9.0.5
Pillow
python-multipart
qdrant-client
python-dotenv
torch==2.7.1+cpu
--extra-index-url https://download.pytorch.org/whl/cpu